Feature/buffered queue (#61)
This commit adds Buffers that transfer the data in chunks. Meaning the reader will see new data only when the writer has filled a chunk, or there was a timeout and a non full chunk was written to the buffer. This greatly accelerates reading and writing to buffers (or queues as they were know before) because now a lock has to be acquired only for each chunk not each item.

The buffers, however, also need to time out to prevent the execution further down the pipeline from stalling. This especially needs to happen when no new items are coming through the pipeline. To implement this, a heartbeat 'tick' was used that is executed in every step along with the normal computation step of each operator. Operators are supposed to pass on their tick to their connectors which in turn pass it on to their buffers (which then might flush) and other operators.
This tick could later also be used for windows.

This commit also adds another, more efficient Connector class that does not use UUIDs and Hashmaps, as they are currently not needed yet.
This commit also adds a Config class that helps to configure the buffers for testing when a tiny timeout is required.